CourseRegistration
CourseRegistration Update

+++ CourseRegistration/Course.cs
@@ -7,8 +7,8 @@ namespace CourseRegistration
 {
     class Course
     {
-        Int32 _ID;
-        public Int32 ID
+        Int16 _ID;
+        public Int16 ID
         {
             get { return _ID; }
             set { _ID = value; }
